VMS-15486: review fixes
VMS-15703: fix crash on remove resource
VMS-15964: storage metrics UT, 'activity' group
VMS-15964: storage metrics UT, 'info' group
VMS-15964: storage metrics UT, 'space' group
VMS-15964: storage metrics UT, 'state' group
VMS-15968: Client has crashed on Windows: c0000005 (Access violation)
VMS-15968: review fixed
VMS-15984: emit signal from p2pMessageBus if remote peer can't accept connection

<==== Mercurial data ====>
Branch: vms_4.1

+++ vms/libs/common/src/core/resource/security_cam_resource.cpp
@@ -1549,8 +1549,8 @@ float QnSecurityCamResource::suggestBitrateKbps(
         auto streamCapability = cameraMediaCapability().streamCapabilities.value(toStreamIndex(role));
         if (streamCapability.maxBitrateKbps > 0)
         {
-            result = qBound((float) streamCapability.minBitrateKbps,
-                result, (float) streamCapability.maxBitrateKbps);
+            result = qBound(streamCapability.minBitrateKbps,
+                result, streamCapability.maxBitrateKbps);
         }
         return result;
     }
